+++ statham/dsl/constants.py
@@ -36,7 +36,6 @@ COMPOSITION_KEYWORDS = ("anyOf", "oneOf", "allOf", "not")
 
 UNSUPPORTED_SCHEMA_KEYWORDS = {
     "$defs",
-    "dependencies",
     "if",
     "then",
     "else",