+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Types of changes are:
 ## [Unreleased]
 ### Added
 * Added support for `minProperties` and `maxProperties` keywords.
+
 ## [0.6.0] - 2020-04-03
 ### Added
 * Added support for `not` keyword.